+++ src/storage/Routes.sol
@@ -31,8 +31,8 @@ library RouteStore {
         uint128 baseFee;
         uint256 relativeGasPriceNumerator;
         uint256 relativeGasPriceDenominator;
-        uint256 gasCoef0;
-        uint256 gasCoef1;
+        uint64 gasCoef0;
+        uint64 gasCoef1;
     }
 
     /**
@@ -51,8 +51,8 @@ library RouteStore {
         uint256 relativeGasPriceDenominator,
         uint128 baseFee,
         uint64 gasLimit,
-        uint256 gasCoef0,
-        uint256 gasCoef1
+        uint64 gasCoef0,
+        uint64 gasCoef1
     );
 
     /**
